# hazelnet-community/src/main/resources/hazelnet-community-openapi.yml
openapi: 3.0.0
info:
  version: '0.0.1'
  title: 'HazelNet Community'
servers:
  - url: 'localhost:8085'
paths:
...
  /externalaccounts/discord/{discordUserId}:
    get:
      summary: Get external account matching a Discord user ID
      operationId: getExternalAccountForDiscordUser
      tags:
        - account
      parameters:
        - $ref: '#/components/parameters/DiscordUserIdPathParameter'
      responses:
        '200':
          description: 'External Account Info that was created or matching existing external account'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExternalAccount'
        '404':
          description: No external account found for this Discord user ID
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    put:
      summary: 'Create/Retrieve External account for Discord user'
      operationId: setExternalAccountForDiscordUser
      tags:
        - account
      parameters:
        - $ref: '#/components/parameters/DiscordUserIdPathParameter'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ExternalAccount'
      responses:
        '200':
          description: 'External Account Info that was created or matching existing external account'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExternalAccount'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /discord/servers:
    get:
      summary: Get all Discord servers
      operationId: listDiscordServers
      tags:
        - discord
      responses:
        '201':
          description: 'List of Discord servers'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DiscordServer'

    post:
      summary: Add Discord Server
      operationId: addDiscordServer
      tags:
        - discord
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DiscordServer'
      responses:
        '201':
          description: 'Discord Server'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiscordServer'
          headers:
            Location:
              $ref: '#/components/schemas/CreatedResponseHeader'
  /discord/servers/{guildId}:
    get:
      summary: Get Discord Server Details
      operationId: getDiscordServer
      tags:
        - discord
      parameters:
        - $ref: '#/components/parameters/GuildIdPathParameter'
      responses:
        '200':
          description: 'Discord Server'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiscordServer'
  /discord/servers/{guildId}/roleassignments/delegatorroles:
    get:
      summary: Get the current set of delegator role assignments for this server
      operationId: getCurrentDelegatorRoleAssignments
      tags:
        - discord
      parameters:
        - $ref: '#/components/parameters/GuildIdPathParameter'
      responses:
        '200':
          description: 'Role assignment list for delegation roles'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DiscordRoleAssignment'
  /discord/servers/{guildId}/roleassignments/tokenroles:
    get:
      summary: Get the current set of token role assignments for this server
      operationId: getCurrentTokenRoleAssignments
      tags:
        - discord
      parameters:
        - $ref: '#/components/parameters/GuildIdPathParameter'
      responses:
        '200':
          description: 'Role assignment list for token roles'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DiscordRoleAssignment'
  /discord/servers/{guildId}/members:
    get:
      summary: Get the list of external accounts that are members of this discord server
      operationId: getExternalAccounts
      tags:
        - discord
      parameters:
        - $ref: '#/components/parameters/GuildIdPathParameter'
      responses:
        '200':
          description: 'Discord Member List'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DiscordMember'
    post:
      summary: Connect an external account to this discord server
      operationId: connectExternalAccount
      tags:
        - discord
      parameters:
        - $ref: '#/components/parameters/GuildIdPathParameter'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DiscordMember'
      responses:
        '201':
          description: 'Added discord member'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiscordMember'
  /discord/servers/{guildId}/members/{externalAccountId}:
    delete:
      summary: Unlink an external account from this discord server
      operationId: disconnectExternalAccount
      tags:
        - discord
      parameters:
        - $ref: '#/components/parameters/GuildIdPathParameter'
        - $ref: '#/components/parameters/ExternalAccountPathParameter'
      responses:
        '204':
          description: 'Successful deletion'
  /discord/servers/{guildId}/tokenpolicies:
    post:
      summary: Add an official token policy to this discord server
      operationId: addTokenPolicy
      tags:
        - discord
        - token
      parameters:
        - $ref: '#/components/parameters/GuildIdPathParameter'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenPolicy'
      responses:
        '201':
          description: 'Token Policy'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenPolicy'
          headers:
            Location:
              $ref: '#/components/schemas/CreatedResponseHeader'
  /discord/servers/{guildId}/tokenpolicies/{policyId}:
    delete:
      summary: Remove an official token policy from this discord server
      operationId: deleteTokenPolicy
      tags:
        - discord
        - token
      parameters:
        - $ref: '#/components/parameters/GuildIdPathParameter'
        - $ref: '#/components/parameters/PolicyIdPathParameter'
      responses:
        '204':
          description: Successful deletion
  /discord/servers/{guildId}/stakepools:
    post:
      summary: Add an official stakepool to this discord server
      operationId: addStakepool
      tags:
        - discord
        - stakepool
      parameters:
        - $ref: '#/components/parameters/GuildIdPathParameter'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Stakepool'
      responses:
        '201':
          description: 'Details of added stakepool'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Stakepool'
          headers:
            Location:
              $ref: '#/components/schemas/CreatedResponseHeader'
  /discord/servers/{guildId}/stakepools/{poolHash}:
    delete:
      summary: Remove an official stakepool from this discord server
      operationId: deleteStakepool
      tags:
        - discord
        - stakepool
      parameters:
        - $ref: '#/components/parameters/GuildIdPathParameter'
        - $ref: '#/components/parameters/PoolHashPathParameter'
      responses:
        '204':
          description: Successful deletion
  /discord/servers/{guildId}/delegatorroles:
    post:
      summary: Add an auto-assigned role based on stakepool delegation to this discord server
      operationId: addDelegatorRole
      tags:
        - discord
        - stakepool
      parameters:
        - $ref: '#/components/parameters/GuildIdPathParameter'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DiscordDelegatorRole'
      responses:
        '201':
          description: 'Token Policy'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiscordDelegatorRole'
          headers:
            Location:
              $ref: '#/components/schemas/CreatedResponseHeader'
  /discord/servers/{guildId}/delegatorroles/{delegatorRoleId}:
    delete:
      summary: Remove a delegator role assignment from this discord server
      operationId: deleteDelegatorRole
      tags:
        - discord
        - stakepool
      parameters:
        - $ref: '#/components/parameters/GuildIdPathParameter'
        - $ref: '#/components/parameters/DelegatorRoleIdPathParameter'
      responses:
        '204':
          description: Successful deletion
  /discord/servers/{guildId}/tokenroles:
    post:
      summary: Add an auto-assigned role based on token ownership to this discord server
      operationId: addTokenOwnershipRole
      tags:
        - discord
        - token
      parameters:
        - $ref: '#/components/parameters/GuildIdPathParameter'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DiscordTokenRole'
      responses:
        '201':
          description: 'Token Policy'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiscordTokenRole'
          headers:
            Location:
              $ref: '#/components/schemas/CreatedResponseHeader'
  /discord/servers/{guildId}/tokenroles/{tokenRoleId}:
    delete:
      summary: Remove a token role assignment from this discord server
      operationId: deleteTokenOwnershipRole
      tags:
        - discord
        - token
      parameters:
        - $ref: '#/components/parameters/GuildIdPathParameter'
        - $ref: '#/components/parameters/TokenRoleIdPathParameter'
      responses:
        '204':
          description: Successful deletion
...
  /discord/servers/{guildId}/whitelists:
    post:
      summary: Add a whitelist for owners of a specific role to this server
      operationId: addWhitelist
      tags:
        - discord
        - whitelist
      parameters:
        - $ref: '#/components/parameters/GuildIdPathParameter'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Whitelist'
      responses:
        '201':
          description: 'Whitelist'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Whitelist'
          headers:
            Location:
              $ref: '#/components/schemas/CreatedResponseHeader'
  /discord/servers/{guildId}/whitelists/{whitelistId}:
    delete:
      summary: Remove a whitelist from this discord server
      operationId: deleteWhitelist
      tags:
        - discord
        - whitelist
      parameters:
        - $ref: '#/components/parameters/GuildIdPathParameter'
        - $ref: '#/components/parameters/WhitelistIdPathParameter'
      responses:
        '204':
          description: Successful deletion
  /discord/servers/{guildId}/whitelists/{whitelistId}/signups:
    get:
      summary: 'Get signups for a whitelist'
      operationId: getWhitelistSignups
      tags:
        - discord
        - whitelist
      parameters:
        - $ref: '#/components/parameters/GuildIdPathParameter'
        - $ref: '#/components/parameters/WhitelistIdOrNamePathParameter'
      responses:
        '200':
          description: 'All signups for this whitelist'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/WhitelistSignup'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Sign up for a whitelist
      operationId: addWhitelistSignup
      tags:
        - discord
        - whitelist
      parameters:
        - $ref: '#/components/parameters/GuildIdPathParameter'
        - $ref: '#/components/parameters/WhitelistIdPathParameter'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WhitelistSignup'
      responses:
        '201':
          description: 'Whitelist signup information'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WhitelistSignup'
          headers:
            Location:
              $ref: '#/components/schemas/CreatedResponseHeader'
  /discord/servers/{guildId}/whitelists/{whitelistId}/signups/{externalAccountId}:
    get:
      summary: Get the whitelist signup for a specific external account for this Discord server
      operationId: getWhitelistSignup
      tags:
        - discord
        - whitelist
      parameters:
        - $ref: '#/components/parameters/GuildIdPathParameter'
        - $ref: '#/components/parameters/WhitelistIdPathParameter'
        - $ref: '#/components/parameters/ExternalAccountPathParameter'
      responses:
        '200':
          description: 'Whitelist signup information'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WhitelistSignup'
        '404':
          description: 'No whitelist signup for this external account present'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    delete:
      summary: Remove a whitelist signup from this discord server
      operationId: deleteWhitelistSignup
      tags:
        - discord
        - whitelist
      parameters:
        - $ref: '#/components/parameters/GuildIdPathParameter'
        - $ref: '#/components/parameters/WhitelistIdPathParameter'
        - $ref: '#/components/parameters/ExternalAccountPathParameter'
      responses:
        '204':
          description: Successful deletion
  /discord/servers/{guildId}/settings/{settingName}:
    put:
      summary: Create or update a guild setting with the provided value
      operationId: updateSetting
      tags:
        - discord
      parameters:
        - $ref: '#/components/parameters/GuildIdPathParameter'
        - $ref: '#/components/parameters/SettingNamePathParameter'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DiscordServerSetting'
      responses:
        '200':
          description: 'Token Policy'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiscordServerSetting'
    delete:
      summary: 'Delete guild setting by Name'
      operationId: deleteSetting
      tags:
        - discord
      parameters:
        - $ref: '#/components/parameters/GuildIdPathParameter'
        - $ref: '#/components/parameters/SettingNamePathParameter'
      responses:
        '204':
          description: 'Successful deletion'
        default:
          description: Unexpected error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /discord/servers/{guildId}/accesstoken:
    post:
      summary: Generate (or regenerate) the API access token for this Discord server for the external HAZELnet API
      operationId: regenerateAccessToken
      tags:
        - discord
        - security
      parameters:
        - $ref: '#/components/parameters/GuildIdPathParameter'
      responses:
        '201':
          description: 'Access token to use with authorization header in external API'
          content:
            application/json:
              schema:
                type: string
          headers:
            Location:
              $ref: '#/components/schemas/CreatedResponseHeader'
    delete:
      summary: Delete the currently active access token for this Discord server
      operationId: deleteAccessToken
      tags:
        - discord
        - security
      parameters:
        - $ref: '#/components/parameters/GuildIdPathParameter'
      responses:
        '204':
          description: Successful deletion
components:
  parameters:
    GuildIdPathParameter:
      name: guildId
      in: path
      description: The Discord Server's Guild ID
      required: true
      style: simple
      schema:
        $ref: '#/components/schemas/DiscordGuild'
    DiscordUserIdPathParameter:
      name: discordUserId
      in: path
      description: The Discord Users's ID
      required: true
      style: simple
      schema:
        $ref: '#/components/schemas/DiscordUser'
    PolicyIdPathParameter:
      name: policyId
      in: path
      description: The Policy ID
      required: true
      style: simple
      schema:
        $ref: '#/components/schemas/TokenPolicyId'
    PoolHashPathParameter:
      name: poolHash
      in: path
      description: The Stakepool hash
      required: true
      style: simple
      schema:
        $ref: '#/components/schemas/StakepoolHash'
    DelegatorRoleIdPathParameter:
      name: delegatorRoleId
      in: path
      description: The ID of this delegator role assignment
      required: true
      style: simple
      example: 235
      schema:
        type: integer
        format: int64
    TokenRoleIdPathParameter:
      name: tokenRoleId
      in: path
      description: The ID of this token role assignment
      required: true
      style: simple
      example: 155
      schema:
        type: integer
        format: int64
...
    WhitelistIdPathParameter:
      name: whitelistId
      in: path
      description: The ID of this whitelist
      required: true
      style: simple
      example: 231
      schema:
        type: integer
        format: int64
...
    SettingNamePathParameter:
      name: settingName
      in: path
      required: true
      style: simple
      example: WHITELISTED_GUILDS
      schema:
        type: string
...
    ExternalAccountPathParameter:
      name: externalAccountId
      in: path
      description: The ID of the external account
      required: true
      style: simple
      schema:
        type: integer
        format: int64
  schemas:
...
    ExternalAccount:
      type: object
      required:
        - type
        - referenceId
      properties:
        id:
          readOnly: true
          type: integer
          format: int64
        registrationTime:
          readOnly: true
          type: string
          format: date-time
          example: 2021-11-10T13:20:57Z
        type:
          $ref: '#/components/schemas/ExternalAccountType'
        referenceId:
          type: string
        referenceName:
          type: string
        account:
          type: integer
          format: int64
...
    DiscordServer:
      type: object
      required:
        - guildId
        - guildName
        - guildOwner
        - guildMemberCount
      properties:
        id:
          readOnly: true
          type: integer
          format: int64
        guildId:
          type: integer
          format: int64
          example: 717264144759390238
        guildName:
          type: string
          example: Guild Name
        guildOwner:
          type: integer
          format: int64
          example: 9358923890
        joinTime:
          readOnly: true
          type: string
          format: date-time
          example: 2021-11-10T13:20:57Z
        guildMemberCount:
          type: integer
          format: int32
          example: 150
        ownerAccount:
          type: integer
          format: int64
        tokenPolicies:
          type: array
          items:
            $ref: '#/components/schemas/TokenPolicy'
        stakepools:
          type: array
          items:
            $ref: '#/components/schemas/Stakepool'
        delegatorRoles:
          type: array
          items:
            $ref: '#/components/schemas/DiscordDelegatorRole'
        tokenRoles:
          type: array
          items:
            $ref: '#/components/schemas/DiscordTokenRole'
        settings:
          type: object
          additionalProperties:
            type: string
    TokenPolicy:
      type: object
      required:
        - policyId
        - projectName
      properties:
        policyId:
          $ref: '#/components/schemas/TokenPolicyId'
        projectName:
          type: string
          example: My First CNFT
    Stakepool:
      type: object
      required:
        - poolHash
      properties:
        poolHash:
          $ref: '#/components/schemas/StakepoolHash'
        info:
          $ref: '#/components/schemas/StakepoolInfo'
...
    DiscordDelegatorRole:
      type: object
      required:
        - minimumStake
        - roleId
      properties:
        id:
          readOnly: true
          type: integer
          format: int64
        poolHash:
          $ref: '#/components/schemas/StakepoolHash'
        minimumStake:
          type: integer
          description: Minimum stake in Lovelace
          format: int64
          example: 1000000000
        roleId:
          $ref: '#/components/schemas/DiscordRole'
    DiscordTokenRole:
      type: object
      required:
        - policyId
        - minimumTokenQuantity
        - roleId
      properties:
        id:
          readOnly: true
          type: integer
          format: int64
        policyId:
          $ref: '#/components/schemas/TokenPolicyId'
        assetFingerprint:
          $ref: '#/components/schemas/AssetFingerprint'
        minimumTokenQuantity:
          type: integer
          format: int64
          example: 1
        roleId:
          $ref: '#/components/schemas/DiscordRole'
...
    Whitelist:
      type: object
      required:
        - displayName
        - name
        - requiredRoleId
      properties:
        id:
          readOnly: true
          type: integer
          format: int64
        displayName:
          type: string
          minLength: 1
          maxLength: 256
        name:
          type: string
          pattern: '^[A-Za-z][-A-Za-z0-9]{0,29}$'
        signupUntil:
          type: string
          format: date-time
          example: 2021-11-10T13:20:57Z
        requiredRoleId:
          $ref: '#/components/schemas/DiscordRole'
...
    DiscordMember:
      type: object
      required:
        - externalAccountId
      properties:
        externalAccountId:
          type: integer
          format: int64
    DiscordRoleAssignment:
      type: object
      required:
        - guildId
        - userId
        - roleId
      properties:
        guildId:
          $ref: '#/components/schemas/DiscordGuild'
        userId:
          $ref: '#/components/schemas/DiscordUser'
        roleId:
          $ref: '#/components/schemas/DiscordRole'
...
    CreatedResponseHeader:
      type: string
      format: uri
      description: URL of the newly created resource
...